sched: Use an enum to tell apart different set_task_rq() callers
set_task_rq() is used for the following four cases:

1. __set_task_cpu()
2. sched_change_group() from fork path
3. sched_change_group() during cgroup move
4. sched_change_group() during autogroup move

Add enum set_task_rq_type which enumerates the above cases and pass it into
set_task_rq() so that it knows what it's being called for. This replaces
TASK_SET_GROUP / TASK_MOVE_GROUP constants which were used by the
->task_change_group() operation.

This doesn't cause any functional changes. The extra parameter will be used
by future changes.

This is pretty messy. Upstream code is being cleaned up in this area, so we
should be able to clean this up in the future too.

Signed-off-by: Tejun Heo <tj@kernel.org>