cgroup: Implement cgroup_show_cftypes()
Implement cgroup_show_cftypes() which show and hide all cgroup files
associated with the specified set of cgroup types. CFTYPE_HIDDEN flag is
added so that files can be created hidden from the get-go.

cgroup_show_cftypes() can be used whether the cftypes are added or not and
combines with cgroup_show_file() so that a given file is visible iff both
its cftype and cfile are visible.

Signed-off-by: Tejun Heo <tj@kernel.org>