sched_ext: Add boilerplate for extensible scheduler class
This introduces dummy implementations of sched_ext interfaces which are used
in scheduler core and hook them in the correct places. As they're all
dummies, this doesn't cause any behavior changes and split out as a separate
patch to help reviewing.

Note that @for_cgroup is added to sched_moved_task() to distinguish cgroup
and autogroup moves. The cgroup moves are notified to sched_ext through
cgroup_move_scx().

Signed-off-by: Tejun Heo <tj@kernel.org>