bpf: Nerf btf scalar value check
scx_bpf_error() calling scx_bpf_error_bstr() fails this check for both fmt
and params parameters. Couldn't find a different way to avoid the issue for
params. For now, nerf the check. Alexei says this part of the code is being
worked on, so hopefully there will be a proper way to support
scx_bpf_error() in the future.

NOT_SIGNED_OFF
Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Dave Marchevsky <davemarchevsky@meta.com>