sched_ext: Add a rust userspace hybrid example scheduler
Atropos is a multi-domain BPF / userspace hybrid scheduler where the BPF
part does simple round robin in each domain and the userspace part
calculates the load factor of each domain and tells the BPF part how to load
balance the domains.

This scheduler demonstrates dividing scheduling logic between BPF and
userspace and using rust to build the userspace part. An earlier variant of
this scheduler was used to balance across six domains, each representing a
chiplet in a six-chiplet AMD processor, and could match the performance of
production setup using CFS.

v3: The userspace code is substantially restructured and rewritten. The
    binary is renamed to scx_atropos and can now figure out the domain
    topology automatically based on L3 cache configuration. The LB logic
    which was rather broken in the previous postings are revamped and should
    behave better.

v2: Updated to use generic BPF cpumask helpers.

Signed-off-by: Dan Schatzberg <dschatzberg@meta.com>
Signed-off-by: Tejun Heo <tj@kernel.org>