scx: Remove unnecessary pr_err on ops.prep_enable() failure path
When ops.prep_enable() returns an error code that is propagated through
scx_ops_prepare_task(), we currently issue a pr_err from
scx_ops_enable(). It seems wrong for a BPF prog to be able to cause the
kernel to issue a pr_err message, and we already capture the error in an
ops_sanitizer_err() call in scx_ops_prepare_task(). Let's remove it.

Signed-off-by: David Vernet <void@manifault.com>