nest: Implement Inria-Paris Nest scheduler
Researchers at Inria Paris recently published a paper called OS
Scheduling with Nest: Keeping Tasks Close Together on Warm Cores [0].
The core idea of the scheduler is to make scheduling decisions which
encourage work to run on cores that are expected to have high frequency.
The scheduler described in their paper is built on top of CFS, and
supports mutli-NUMA and multi-CCX architectures. This patch implements
a more basic version of the Nest scheduler which should work on
single-CCX or single-socket machines.

A subsequent patch will improve logging for the scheduler, and in the
future we should look into making this support multi-CCX/NUMA. Doing so
would likely involve integrating with scx_rusty, or leveraging its user
space load balancer as a pluggable component.

[0]: https://hal.inria.fr/hal-03612592/file/paper.pdf

Signed-off-by: David Vernet <void@manifault.com>