scx_nest: Rebase onto latest and add -I flag for checking hypertwins
The default behavior of the original Inria Nest scheduler is to ignore
hypertwins and select any idle core in a Nest. Let's make the default
behavior match the original Nest scheduler by making the default
behavior to ignore hypertwins. If the -I flag is passed, the scheduler
will first check for fully idle cores.

Signed-off-by: David Vernet <void@manifault.com>