scx_nest: Reset prev_misses when r_impatient is exceeded
We weren't resetting tctx->prev_misses when a task failed to get its
previous core r_impatient times in a row. This could cause us to
incorrectly skip looking for primary nest cores when the value is
exceeded if we aren't able to use our previous core on subsequent
wakeups.

In addition, clean the code up a bit.

Signed-off-by: David Vernet <void@manifault.com>