scx_nest: Compact based on repeated comapction schedulings
When deciding whether to compact a nest, we're currently looking at how
much work was scheduled on the core before it's scheduled for a
compaction. This is not a great solution though. The whole point of nest
is that we're trying to keep a core warm on the assumption of temporal
locality, and that it will be used again soon. What would be preferable
is to aggressively compact it if it's repeatedly being scheduled for
compaction, as it should only be repeatedly scheduled for compaction if
there aren't enough tasks to fully utilize the primary nest.

Signed-off-by: David Vernet <void@manifault.com>