sched_ext: introduce SCX_OPS_PM_AUTO_EXIT
A sched_ext scheduler may have dependencies on external user-space
components. In presence of a power-management (PM) event, such as
suspend or hibernate, these user-space components can be frozen,
potentially preventing any further progress of the scheduler.

This leads to the active scheduler (scx_userland in the following
example) being disabled, reporting an oops like the following in the
kernel log, that could be misinterpreted as an actual stall:

 sched_ext: BPF scheduler "userland" errored, disabling
 sched_ext: runnable task stall (watchdog failed to check in for 3.565s)
    scheduler_tick+0x1d7/0x3f0
    update_process_times+0x89/0xb0
    tick_sched_handle+0x28/0x70
    tick_nohz_highres_handler+0x78/0xa0
    __hrtimer_run_queues+0x10f/0x2a0
    hrtimer_interrupt+0xf6/0x250
    __sysvec_apic_timer_interrupt+0x4e/0x150
    sysvec_apic_timer_interrupt+0x8d/0xd0
    asm_sysvec_apic_timer_interrupt+0x1b/0x20
    cpuidle_enter_state+0xda/0x730
    cpuidle_enter+0x2e/0x50
    call_cpuidle+0x23/0x60
    cpuidle_idle_call+0x11d/0x190
    do_idle+0x87/0xf0
    cpu_startup_entry+0x2a/0x30
    start_secondary+0x129/0x160
    secondary_startup_64_no_verify+0x17d/0x18b

While deactivating the sched-ext scheduler on a suspend/resume cycle is
a reasonable approach for now, a more elegant solution would be to
report an explicit message indicating that the scheduler has been
disabled due to a PM suspend event.

To address this issue, introduce the SCX_OPS_PM_AUTO_EXIT flag, which
can be optionally set to automatically disable the scheduler when any PM
event occurs.

With this flag set (in sched_ext_ops.flags), the scheduler is
automatically disabled on a PM event and the message reported in the
kernel log looks like this:

 sched_ext: "userland" disabled on PM event

Subsequently, the scheduler can be easily reactivated on resume, if
necessary, through user-space actions, for example using a
pm-utils/systemd script or similar solutions.

Schedulers that do not depend on user-space components are not affected
by this issue (because jiffies are not progressing while the system is
suspended / hibernated), so they can ignore the SCX_OPS_PM_AUTO_EXIT
flag.

This can help to identify potential "false positive" stalls in systems
where PM events are frequent, such as laptops or mobile / embedded
devices.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>