scx: Annotate kfuncs and support CFI
The most recent struct bpf_struct_ops implementation has a cfi_stubs
field so as to enable CFI calls for indirect struct_ops callbacks. We
have to add a bunch of blank callbacks so we generate the correct CFI
hash and let the compiler generate the correct CFI for the actual
trampolines.

In addition, use the __bpf_kfunc annotation on scx kfuncs.

Signed-off-by: David Vernet <void@manifault.com>