scx: Always keep tasks on dsq->list
When a DSQ is used as a PRIQ, tasks were only kept on dsq->priq. This
requires all other paths to check both dsq->priq and dsq->list. A task's
scx.dsq_node.list is available when the task is on a PRIQ DSQ anyway, let's
always put tasks on dsq->list whether it's used as a PRIQ or FIFO by
inserting into dsq->list in vtime order on PRIQs.

This means that dsq->list is always in dispatch order whether the DSQ is
FIFO or PRIQ and the only paths that need to worry about the DSQ type are
dispatch_enqueue() and task_unlink_from_dsq().