scx: Fix CPU online tracking
Commit 10d167d8ab81 ("SCX: Synchronize boot CPU's online state with
SCX_RQ_ONLINE") fixes the oversight that we're not synchronizing the boot CPU's
online state with scx_rq's online state. It does this by adding a check in
init_sched_ext_class() for whether a CPU is online, and setting the
SCX_RQ_ONLINE flag in the rq if it is.

The online mask may not be initialized by the time init_sched_ext_class() is
invoked, so let's add a late init function where we iterate over the possible
CPUs and do this check there.

Signed-off-by: David Vernet <void@manifault.com>