scx: Check for NULLity for ops.yield() @to arg
The @to field in ops.yield() can be NULL (when the callback is invoked)
on the yield_task_scx() path). Right now we're not communicating to the
verifier that the arg can be NULL, so if callers don't check for NULL it
can crash the host. Let's update ext.c to check. A subsequent patch will
add tests.

Signed-off-by: David Vernet <void@manifault.com>